+++ macOS/DuckDuckGo/SetAsDefaultAddToDockPrompts/BannerView.swift
